avoid warning on comparison of signed and unsigned int
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-spgrid/trunk@146 f45f52a6-3457-4210-b8e0-e9d642ddcde3

+++ dune/grid/spgrid/indexset.hh
@@ -307,7 +307,7 @@ namespace Dune
     const LevelIndexSet &levelIndexSet ( const int level ) const
     {
       assert( (level >= 0) && (level <= (int)levelIndexSets_.size()) );
-      assert( levelIndexSets_.size() == grid().maxLevel()+1 );
+      assert( (int)levelIndexSets_.size() == grid().maxLevel()+1 );
       return *levelIndexSets_[ level ];
     }
